+++ HealthWebApp/HealthWebApp/Data/EntityModel/Barangays/BarangayOfficial.cs
@@ -4,10 +4,10 @@ namespace HealthWebApp.Data.EntityModel.Barangays
 {
     public class BarangayOfficial
     {
-        public int Id { get; set; }
-        public int BarangayId { get; set; }
+        public long Id { get; set; }
+        public long BarangayId { get; set; }
         public virtual Barangay Barangay { get; set; }
-        public int PersonId { get; set; }
+        public long PersonId { get; set; }
         public virtual Person Officer { get; set; }
         public int DesignationId { get; set; }
         public virtual Designation Position { get; set; }
