+++ HealthWebApp/HealthWebApp/Data/EntityModel/OutPatient/DoctorsOrder.cs
@@ -20,9 +20,9 @@ namespace HealthWebApp.Data.EntityModel.OutPatient
 
     public class DoctorsOrder //pg109
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int ConsultationId { get; set; }
+        public long ConsultationId { get; set; }
         public virtual Consultation Consultation { get; set; }
 
         //Laboratory Request
