+++ HealthWebApp/HealthWebApp/Data/EntityModel/OutPatient/RiskAssesment.cs
@@ -7,9 +7,9 @@ namespace HealthWebApp.Data.EntityModel.OutPatient
 {
     public class RiskAssesment
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int ConsultationId { get; set; }
+        public long ConsultationId { get; set; }
         public virtual Consultation Consultation { get; set; }
     }
 }
