+++ HealthWebApp/HealthWebApp/Models/HouseholdProfile/HouseholdProfileCreateModel.cs
@@ -4,7 +4,7 @@
     {
         public string ProfileId { get; set; }
         public string Address { get; set; }
-        public int BarangayId { get; set; }
+        public long BarangayId { get; set; }
         public string Note { get; set; }
     }
 }
