+++ HealthWebApp/HealthWebApp/Models/HouseholdProfile/HouseholdProfileEditModel.cs
@@ -7,10 +7,10 @@ namespace HealthWebApp.Models.HouseholdProfile
 {
     public class HouseholdProfileEditModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string ProfileId { get; set; }
         public string Address { get; set; }
-        public int BarangayId { get; set; }
+        public long BarangayId { get; set; }
         public string Note { get; set; }
 
     }
